requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
openai>=1.12.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
Pillow>=10.0.0
